Succesfully connected to local test db
Website can connect with local mysql server. However, nothing is shown as of yet. Simply a blank login page is present. I had to tweak the database_connection, using the updated mysqli commands. I also had to tweak some of the sql files to create the databases and add user account in phpmyadmin. More details are found in: notes on db.txt