Working admin_videos page. Added 'vids' table to database
Notes for Aaron: in order for the new video admin page to work for you, you will have to add a new table with pre-filled values to the uwalc2 database.  Please refer to the admin/videoInstructionsPics/vids.sql file. I exported that file from my database. So in theory you should be able to simply import it into yours to get it working.