Sidebar admin page no longer redirects
Improved the design of the sidebar admin page by eliminating the redirect page (where then you would have to navigate back to the original page). After image or text change, admin will remain on the same page and get a notification at the top of the page of the result of the change.